ARM: dts: bcm2710 - local_intc users need polarity
Someone upstream decided it was vital that interrupt signals internal
to the BCM283x CPUs had a defined polarity, even though we don't know
what it is. As a result the local_intc node now demands two-cell
declarations and all users have to change to match or risk a warning
from the DTC (at least).

For what it's worth, LEVEL_HIGH was chosen.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>